services/auth/requirements.txt
pydantic==2.5.3
PyJWT==2.8.0
